Error Handling

+++ tftp_client.py
@@ -1,5 +1,5 @@
 import sys
-from tftp_funcs import upload_file, download_file
+from tftp_funcs import upload_file, download_file, check_local_file_exists
 
 ###### Main Program
 def main():
@@ -19,15 +19,18 @@ def main():
         if action == "1": 
             # Store file in server
             file_to_upload = input("Enter the path of the file to upload: ")
+            if not check_local_file_exists(file_to_upload):
+                continue # Skip to the next iteration
             remote_filename = input("Enter the filename to use on the server: ")
-            block_size = int(input("Enter block size for upload (default 512): "))
+            block_size = int(input("Enter block size for upload (default 512): ") or 512)
             upload_file(server_ip, file_to_upload, remote_filename, block_size)
+            continue
 
         elif action == "2":
             # Retrieve file from server
             remote_filename = input("Enter the filename to download: ")
             local_filename = input("Enter the local filename to save the downloaded file: ")
-            block_size = int(input("Enter block size for download (default 512): "))
+            block_size = int(input("Enter block size for download (default 512): ") or 512)
             download_file(server_ip, remote_filename, local_filename, block_size)
 
         elif action == "3":
